Split counterexample build and check
