+++ liquid-fixpoint.cabal
@@ -56,6 +56,9 @@ library
   exposed-modules:  Data.ShareMap
                     Language.Fixpoint.Conditional.Z3
                     Language.Fixpoint.CounterExample
+                    Language.Fixpoint.CounterExample.Build
+                    Language.Fixpoint.CounterExample.Check
+                    Language.Fixpoint.CounterExample.Types
                     Language.Fixpoint.Defunctionalize
                     Language.Fixpoint.Graph
                     Language.Fixpoint.Graph.Deps
